beforeEditHandler: disregard re-dispatches from FlexFormPlugin
FlexFormPlugin dispatches beforeEditHandler for nefarious reasons,
leading to two dispatches within any /bin/edit request that uses
RENDERFOREDIT anywhere. If the data form is rendered based on values in
a template, the second dispatch will refer to the template topic, thus
making KVPPlugin check workflow restrictions for the template, and
apparently getting it quite wrong and bailing out.

As a workaround, only actually perform the check in the first dispatch,
and ignore all subsequent dispatches in the same session.

Ticket: #5049