Also use epoch transition time for indexing
We now store the epoch time of a transition, however only the rounded
value was used for indexing. Using the epoch time when available and
fall back when the transition occured before we stored that time.

Ticket: #6022
Release-Note(feature,appdev) {
KVPPlugin now additionaly stores the epoch time of a transition.
}