ID:8570 Hotfix: avoid approved talks
id: 8570
When mandatory formfields are missing in an approved topic a discussion
remains in the approved state, though the TALK file is created. Ignoring
the mandatory fields on FORK allows proper states and forces the next
editor to fix missing form fields. The fix also provides a
MaintenanceCheck to discover broken Talks but no automated repair.